Add school information to Exam business model. Remove redundant school information from exam history business model.
