+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/jdbc/Exams.java
@@ -2,6 +2,7 @@ package org.opentestsystem.rdw.reporting.common.jdbc;
 
 import org.opentestsystem.rdw.common.model.AssessmentType;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
+import org.opentestsystem.rdw.reporting.common.model.Organization;
 import org.opentestsystem.rdw.reporting.common.model.ScaleScore;
 
 import java.sql.ResultSet;
@@ -30,6 +31,8 @@ public final class Exams {
 
     public static <A extends Exam.Builder> A map(final ResultSet row, final A builder) throws SQLException {
 
+        final Organization school = new Organization(row.getLong("school_id"), row.getString("school_name"));
+
         builder.id(row.getLong("id"))
                 .sessionId(row.getString("session_id"))
                 .dateTime(row.getTimestamp("completed_at").toInstant())
@@ -38,13 +41,14 @@ public final class Exams {
                 .administrativeConditionCode(row.getString("administration_condition_code"))
                 .schoolYear(row.getInt("school_year"))
                 .accommodationCodes(safeSplit(row.getString("available_accommodation_codes"), "|"))
-                .scaleScore(ScaleScores.map(row));
+                .scaleScore(ScaleScores.map(row))
+                .school(school);
 
         // Collects claim scores for ICA and Summative exams.
         // If a single claim is unscored a null entry will be put in the list to represent the unscored claim
         if (AssessmentType.valueOf(row.getInt("type_id")) != AssessmentType.IAB) {
             final List<ScaleScore> claimScaleScores = newArrayList();
-            for (ClaimScaleScoreColumn column : CLAIM_SCALE_SCORE_COLUMNS) {
+            for (final ClaimScaleScoreColumn column : CLAIM_SCALE_SCORE_COLUMNS) {
 
                 final String claimCode = getNullable(row, row.getString(column.code));
                 if (claimCode == null) {
