+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/DistrictGroup.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.organization.model;
+package org.opentestsystem.rdw.reporting.common.model;
 
 /**
  * Represents a group of districts.
