+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/Exam.java
@@ -23,6 +23,7 @@ public class Exam {
     private ScaleScore scaleScore;
     private List<ScaleScore> claimScaleScores;
     private List<String> accommodationCodes;
+    private Organization school;
 
     private Exam() {
     }
@@ -114,6 +115,13 @@ public class Exam {
         return accommodationCodes;
     }
 
+    /**
+     * @return The school this exam was taken at
+     */
+    public Organization getSchool() {
+        return school;
+    }
+
     public static Builder builder() {
         return new Builder();
     }
@@ -132,6 +140,7 @@ public class Exam {
         private String administrativeConditionCode;
         private List<ScaleScore> claimScaleScores;
         private List<String> accommodationCodes;
+        private Organization school;
 
         public Builder id(final long id) {
             this.id = id;
@@ -193,6 +202,11 @@ public class Exam {
             return this;
         }
 
+        public Builder school(final Organization school) {
+            this.school = school;
+            return this;
+        }
+
         public Exam build() {
             final Exam exam = new Exam();
             exam.id = id;
@@ -207,6 +221,7 @@ public class Exam {
             exam.administrativeConditionCode = administrativeConditionCode;
             exam.claimScaleScores = claimScaleScores != null ? Collections.unmodifiableList(claimScaleScores) : null;
             exam.accommodationCodes = accommodationCodes != null ? ImmutableList.copyOf(accommodationCodes) : ImmutableList.of();
+            exam.school = school;
             return exam;
         }
