+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/School.java
@@ -1,14 +1,22 @@
-package org.opentestsystem.rdw.reporting.organization.model;
+package org.opentestsystem.rdw.reporting.common.model;
 
 /**
- * Represents a group of schools.
+ * Represents a school
  */
-public class SchoolGroup extends Organization {
+public class School extends Organization {
+
+    private Long schoolGroupId;
     private long districtId;
     private Long districtGroupId;
-
-    SchoolGroup(final long id, final String name) {
+    School(final long id, final String name) {
         super(id, name);
+    }
+
+    /**
+     * @return The school group id that this school belongs to.
+     */
+    public Long getSchoolGroupId() {
+        return schoolGroupId;
     }
 
     /**
@@ -25,14 +33,12 @@ public class SchoolGroup extends Organization {
         return districtGroupId;
     }
 
-    public static Builder builder() {
-        return new Builder();
-    }
+    public static Builder builder() { return new Builder(); }
 
     public static class Builder {
-
         private long id;
         private String name;
+        private Long schoolGroupId;
         private long districtId;
         private Long districtGroupId;
 
@@ -46,6 +52,11 @@ public class SchoolGroup extends Organization {
             return this;
         }
 
+        public Builder schoolGroupId(final Long schoolGroupId) {
+            this.schoolGroupId = schoolGroupId;
+            return this;
+        }
+
         public Builder districtId(final long districtId) {
             this.districtId = districtId;
             return this;
@@ -56,13 +67,15 @@ public class SchoolGroup extends Organization {
             return this;
         }
 
-        public SchoolGroup build() {
-            final SchoolGroup group = new SchoolGroup(id, name);
+        public School build() {
+            final School school = new School(this.id, this.name);
 
-            group.districtId = districtId;
-            group.districtGroupId = districtGroupId;
+            school.schoolGroupId = schoolGroupId;
+            school.districtId = districtId;
+            school.districtGroupId = districtGroupId;
 
-            return group;
+            return school;
         }
     }
+
 }
