+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/SchoolGroup.java
@@ -1,22 +1,14 @@
-package org.opentestsystem.rdw.reporting.organization.model;
+package org.opentestsystem.rdw.reporting.common.model;
 
 /**
- * Represents a school
+ * Represents a group of schools.
  */
-public class School extends Organization {
-
-    private Long schoolGroupId;
+public class SchoolGroup extends Organization {
     private long districtId;
     private Long districtGroupId;
-    School(final long id, final String name) {
-        super(id, name);
-    }
 
-    /**
-     * @return The school group id that this school belongs to.
-     */
-    public Long getSchoolGroupId() {
-        return schoolGroupId;
+    SchoolGroup(final long id, final String name) {
+        super(id, name);
     }
 
     /**
@@ -33,12 +25,14 @@ public class School extends Organization {
         return districtGroupId;
     }
 
-    public static Builder builder() { return new Builder(); }
+    public static Builder builder() {
+        return new Builder();
+    }
 
     public static class Builder {
+
         private long id;
         private String name;
-        private Long schoolGroupId;
         private long districtId;
         private Long districtGroupId;
 
@@ -52,11 +46,6 @@ public class School extends Organization {
             return this;
         }
 
-        public Builder schoolGroupId(final Long schoolGroupId) {
-            this.schoolGroupId = schoolGroupId;
-            return this;
-        }
-
         public Builder districtId(final long districtId) {
             this.districtId = districtId;
             return this;
@@ -67,15 +56,13 @@ public class School extends Organization {
             return this;
         }
 
-        public School build() {
-            final School school = new School(this.id, this.name);
+        public SchoolGroup build() {
+            final SchoolGroup group = new SchoolGroup(id, name);
 
-            school.schoolGroupId = schoolGroupId;
-            school.districtId = districtId;
-            school.districtGroupId = districtGroupId;
+            group.districtId = districtId;
+            group.districtGroupId = districtGroupId;
 
-            return school;
+            return group;
         }
     }
-
 }
