+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/repository/JdbcIabReportRepositoryIT.java
@@ -8,6 +8,7 @@ import org.opentestsystem.rdw.common.model.AssessmentType;
 import org.opentestsystem.rdw.common.model.Subject;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
+import org.opentestsystem.rdw.reporting.common.model.Organization;
 import org.opentestsystem.rdw.reporting.common.model.ScaleScore;
 import org.opentestsystem.rdw.reporting.common.model.Student;
 import org.opentestsystem.rdw.reporting.common.report.UserPermissions;
@@ -191,6 +192,7 @@ public class JdbcIabReportRepositoryIT {
                 .completenessCode("Complete")
                 .scaleScore(ScaleScore.builder().level(2).value(2100).standardError(21.0).build())
                 .claimScaleScores(null)
+                .school(new Organization(-10, "school1"))
                 .build();
 
         final Exam iab1_exam2 = Exam.builder()
@@ -203,6 +205,7 @@ public class JdbcIabReportRepositoryIT {
                 .completenessCode("Complete")
                 .scaleScore(ScaleScore.builder().level(2).value(2100).standardError(21.0).build())
                 .claimScaleScores(null)
+                .school(new Organization(-10, "school1"))
                 .build();
 
         final Exam iab2_exam1 = Exam.builder()
@@ -215,6 +218,7 @@ public class JdbcIabReportRepositoryIT {
                 .completenessCode("Complete")
                 .scaleScore(ScaleScore.builder().level(2).value(2100).standardError(21.0).build())
                 .claimScaleScores(null)
+                .school(new Organization(-10, "school1"))
                 .build();
 
         final Exam iab2_exam2 = Exam.builder()
@@ -227,6 +231,7 @@ public class JdbcIabReportRepositoryIT {
                 .completenessCode("Complete")
                 .scaleScore(ScaleScore.builder().level(2).value(2100).standardError(21.0).build())
                 .claimScaleScores(null)
+                .school(new Organization(-10, "school1"))
                 .build();
 
         final Map<Assessment, Collection<Exam>> assessments = newTreeMap(comparing(Assessment::getName));
