+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/repository/JdbcIcaReportRepositoryIT.java
@@ -9,6 +9,7 @@ import org.opentestsystem.rdw.common.model.AssessmentType;
 import org.opentestsystem.rdw.common.model.Subject;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
+import org.opentestsystem.rdw.reporting.common.model.Organization;
 import org.opentestsystem.rdw.reporting.common.model.ScaleScore;
 import org.opentestsystem.rdw.reporting.common.model.Student;
 import org.opentestsystem.rdw.reporting.common.report.UserPermissions;
@@ -201,6 +202,7 @@ public class JdbcIcaReportRepositoryIT {
                                                 ScaleScore.builder().level(3).value(300).standardError(30).build(),
                                                 ScaleScore.builder().level(4).value(400).standardError(40).build()
                                         ))
+                                        .school(new Organization(-10, "school1"))
                                         .build()
                                 )
                                 .build()
