+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/organization/DefaultOrganizationService.java
@@ -1,10 +1,10 @@
 package org.opentestsystem.rdw.reporting.organization;
 
 import com.google.common.collect.ImmutableSet;
-import org.opentestsystem.rdw.reporting.organization.model.District;
-import org.opentestsystem.rdw.reporting.organization.model.DistrictGroup;
-import org.opentestsystem.rdw.reporting.organization.model.School;
-import org.opentestsystem.rdw.reporting.organization.model.SchoolGroup;
+import org.opentestsystem.rdw.reporting.common.model.District;
+import org.opentestsystem.rdw.reporting.common.model.DistrictGroup;
+import org.opentestsystem.rdw.reporting.common.model.School;
+import org.opentestsystem.rdw.reporting.common.model.SchoolGroup;
 import org.opentestsystem.rdw.reporting.common.security.User;
 import org.opentestsystem.rdw.security.Permission;
 import org.springframework.beans.factory.annotation.Autowired;
