+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/organization/JdbcOrganizationRepository.java
@@ -1,10 +1,10 @@
 package org.opentestsystem.rdw.reporting.organization;
 
 import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
-import org.opentestsystem.rdw.reporting.organization.model.District;
-import org.opentestsystem.rdw.reporting.organization.model.DistrictGroup;
-import org.opentestsystem.rdw.reporting.organization.model.School;
-import org.opentestsystem.rdw.reporting.organization.model.SchoolGroup;
+import org.opentestsystem.rdw.reporting.common.model.District;
+import org.opentestsystem.rdw.reporting.common.model.DistrictGroup;
+import org.opentestsystem.rdw.reporting.common.model.School;
+import org.opentestsystem.rdw.reporting.common.model.SchoolGroup;
 import org.opentestsystem.rdw.security.PermissionScope;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
