+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/organization/OrganizationRepository.java
@@ -1,9 +1,9 @@
 package org.opentestsystem.rdw.reporting.organization;
 
-import org.opentestsystem.rdw.reporting.organization.model.District;
-import org.opentestsystem.rdw.reporting.organization.model.DistrictGroup;
-import org.opentestsystem.rdw.reporting.organization.model.School;
-import org.opentestsystem.rdw.reporting.organization.model.SchoolGroup;
+import org.opentestsystem.rdw.reporting.common.model.District;
+import org.opentestsystem.rdw.reporting.common.model.DistrictGroup;
+import org.opentestsystem.rdw.reporting.common.model.School;
+import org.opentestsystem.rdw.reporting.common.model.SchoolGroup;
 import org.opentestsystem.rdw.security.PermissionScope;
 
 import javax.validation.constraints.NotNull;
