+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/organization/OrganizationService.java
@@ -1,9 +1,9 @@
 package org.opentestsystem.rdw.reporting.organization;
 
-import org.opentestsystem.rdw.reporting.organization.model.District;
-import org.opentestsystem.rdw.reporting.organization.model.DistrictGroup;
-import org.opentestsystem.rdw.reporting.organization.model.School;
-import org.opentestsystem.rdw.reporting.organization.model.SchoolGroup;
+import org.opentestsystem.rdw.reporting.common.model.District;
+import org.opentestsystem.rdw.reporting.common.model.DistrictGroup;
+import org.opentestsystem.rdw.reporting.common.model.School;
+import org.opentestsystem.rdw.reporting.common.model.SchoolGroup;
 import org.opentestsystem.rdw.reporting.common.security.User;
 import org.springframework.security.access.prepost.PreAuthorize;
