+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/search/exam/student/JdbcStudentExamRepository.java
@@ -6,8 +6,8 @@ import org.opentestsystem.rdw.reporting.common.jdbc.Exams;
 import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
+import org.opentestsystem.rdw.reporting.common.model.Organization;
 import org.opentestsystem.rdw.reporting.common.model.Student;
-import org.opentestsystem.rdw.reporting.organization.model.Organization;
 import org.opentestsystem.rdw.security.Permission;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -72,7 +72,6 @@ class JdbcStudentExamRepository implements StudentExamRepository {
                     return new StudentHistoryExamWrapper.Builder()
                             .exam(exam)
                             .assessment(assessment)
-                            .school(school)
                             .build();
                 }
         );
