+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/search/exam/student/StudentHistoryExamWrapper.java
@@ -2,7 +2,6 @@ package org.opentestsystem.rdw.reporting.search.exam.student;
 
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
-import org.opentestsystem.rdw.reporting.organization.model.Organization;
 
 /**
  * This model represents an Exam within the context of a student's history.
@@ -12,7 +11,6 @@ public class StudentHistoryExamWrapper {
 
     private Exam exam;
     private Assessment assessment;
-    private Organization school;
 
     private StudentHistoryExamWrapper() {
     }
@@ -31,24 +29,15 @@ public class StudentHistoryExamWrapper {
         return assessment;
     }
 
-    /**
-     * @return The exam's School
-     */
-    public Organization getSchool() {
-        return school;
-    }
-
     public static class Builder {
 
         private Exam exam;
         private Assessment assessment;
-        private Organization school;
 
         public StudentHistoryExamWrapper build() {
             final StudentHistoryExamWrapper examWrapper = new StudentHistoryExamWrapper();
             examWrapper.exam = exam;
             examWrapper.assessment = assessment;
-            examWrapper.school = school;
 
             return examWrapper;
         }
@@ -60,11 +49,6 @@ public class StudentHistoryExamWrapper {
 
         public Builder assessment(final Assessment assessment) {
             this.assessment = assessment;
-            return this;
-        }
-
-        public Builder school(final Organization school) {
-            this.school = school;
             return this;
         }
     }
