+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/search/user/UserContext.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.reporting.search.user;
 
 import com.google.common.collect.ImmutableSet;
-import org.opentestsystem.rdw.reporting.organization.model.School;
+import org.opentestsystem.rdw.reporting.common.model.School;
 import org.opentestsystem.rdw.reporting.search.group.Group;
 
 import java.util.Set;
