+++ webapp/src/main/webapp/src/app/assessments/model/exam.model.ts
@@ -1,5 +1,6 @@
 import { ClaimScore } from "./claim-score.model";
 import { Student } from "../../student/model/student.model";
+import { School } from "../../user/model/school.model";
 
 export class Exam {
   id: number;
@@ -20,4 +21,5 @@ export class Exam {
   schoolYear: number;
   claimScores: ClaimScore[];
   accommodationCodes: string[];
+  school: School;
 }
