+++ webapp/src/main/webapp/src/app/student/results/tables/student-history-ica-summitive-table.component.html
@@ -34,7 +34,7 @@
       {{ exam.schoolYear | schoolYear }}
     </ng-template>
   </p-column>
-  <p-column field="school.name" header="{{'labels.student.results.cols.school' | translate }}" sortable="true"></p-column>
+  <p-column field="exam.school.name" header="{{'labels.student.results.cols.school' | translate }}" sortable="true"></p-column>
   <p-column field="exam.enrolledGrade" header="{{'labels.student.results.cols.enrolled-grade' | translate}}" sortable="true">
     <ng-template let-exam="rowData.exam" pTemplate="body">
       {{ exam.enrolledGrade | gradeDisplay }}
