+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/TestData.java
@@ -8,6 +8,7 @@ import org.opentestsystem.rdw.common.model.AssessmentType;
 import org.opentestsystem.rdw.common.model.Subject;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
+import org.opentestsystem.rdw.reporting.common.model.Organization;
 import org.opentestsystem.rdw.reporting.common.model.ScaleScore;
 import org.opentestsystem.rdw.reporting.common.model.Student;
 import org.opentestsystem.rdw.reporting.common.model.StudentContext;
@@ -150,7 +151,8 @@ public final class TestData {
                         ScaleScore.builder().level(2).value(200).standardError(20).build(),
                         ScaleScore.builder().level(3).value(300).standardError(30).build(),
                         ScaleScore.builder().level(4).value(400).standardError(40).build()
-                ));
+                ))
+                .school(new Organization(-10, "school1"));
     }
 
     public static Exam.Builder exam2() {
@@ -163,7 +165,8 @@ public final class TestData {
                 .scaleScore(
                         ScaleScore.builder().level(2).value(2100).standardError(21).build()
                 )
-                .claimScaleScores(null);
+                .claimScaleScores(null)
+                .school(new Organization(-10, "school1"));
     }
 
     public static Exam.Builder exam3() {
@@ -180,7 +183,8 @@ public final class TestData {
                         ScaleScore.builder().level(1).value(1000).standardError(100).build(),
                         ScaleScore.builder().level(2).value(2000).standardError(200).build(),
                         ScaleScore.builder().level(3).value(3000).standardError(300).build()
-                ));
+                ))
+                .school(new Organization(-20, "school2"));
     }
 
     public static Exam.Builder exam4() {
@@ -196,7 +200,8 @@ public final class TestData {
                         ScaleScore.builder().level(1).value(1100).standardError(110).build(),
                         ScaleScore.builder().level(2).value(2100).standardError(210).build(),
                         ScaleScore.builder().level(3).value(3100).standardError(310).build()
-                ));
+                ))
+                .school(new Organization(-30, "school3"));
     }
 
     public static Exam.Builder exam5() {
@@ -212,7 +217,8 @@ public final class TestData {
                         ScaleScore.builder().level(1).value(1200).standardError(120).build(),
                         ScaleScore.builder().level(2).value(2200).standardError(220).build(),
                         ScaleScore.builder().level(3).value(3200).standardError(320).build()
-                ));
+                ))
+                .school(new Organization(-30, "school3"));
     }
 
     public static Exam.Builder exam6() {
@@ -224,7 +230,8 @@ public final class TestData {
                 .administrativeConditionCode("Standardized")
                 .completenessCode("Partial")
                 .scaleScore(null)
-                .claimScaleScores(newArrayList(null, null, null));
+                .claimScaleScores(newArrayList(null, null, null))
+                .school(new Organization(-30, "school3"));
     }
 
 }
