+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/organization/JdbcOrganizationRepositoryIT.java
@@ -2,12 +2,12 @@ package org.opentestsystem.rdw.reporting.organization;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.opentestsystem.rdw.reporting.common.model.District;
+import org.opentestsystem.rdw.reporting.common.model.DistrictGroup;
+import org.opentestsystem.rdw.reporting.common.model.School;
+import org.opentestsystem.rdw.reporting.common.model.SchoolGroup;
 import org.opentestsystem.rdw.reporting.common.test.ITDataSourceConfiguration;
 import org.opentestsystem.rdw.reporting.common.test.RepositoryIT;
-import org.opentestsystem.rdw.reporting.organization.model.District;
-import org.opentestsystem.rdw.reporting.organization.model.DistrictGroup;
-import org.opentestsystem.rdw.reporting.organization.model.School;
-import org.opentestsystem.rdw.reporting.organization.model.SchoolGroup;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
@@ -123,7 +123,7 @@ public class JdbcOrganizationRepositoryIT {
 
     @Test
     public void itShouldReturnExactSchoolsForSchoolsPermissionScope() throws Exception {
-        List<School> schools = repository.findAllSchools(schools(-40L));
+        final List<School> schools = repository.findAllSchools(schools(-40L));
 
         assertThat(schools)
                 .usingRecursiveFieldByFieldElementComparator()
@@ -157,7 +157,7 @@ public class JdbcOrganizationRepositoryIT {
 
     @Test
     public void itShouldReturnExactSchoolsGroupForSchoolsPermissionScope() throws Exception {
-        List<SchoolGroup> districts = repository.findAllSchoolGroups(schools(-40L));
+        final List<SchoolGroup> districts = repository.findAllSchoolGroups(schools(-40L));
 
         assertThat(districts)
                 .usingRecursiveFieldByFieldElementComparator()
@@ -191,7 +191,7 @@ public class JdbcOrganizationRepositoryIT {
 
     @Test
     public void itShouldReturnExactSchoolsDistrictForSchoolsPermissionScope() throws Exception {
-        List<District> districts = repository.findAllDistricts(schools(-40L));
+        final List<District> districts = repository.findAllDistricts(schools(-40L));
 
         assertThat(districts)
                 .usingRecursiveFieldByFieldElementComparator()
@@ -226,7 +226,7 @@ public class JdbcOrganizationRepositoryIT {
 
     @Test
     public void itShouldReturnExactDistrictsGroupForSchoolsPermissionScope() throws Exception {
-        List<DistrictGroup> districts = repository.findAllDistrictGroups(schools(-40L));
+        final List<DistrictGroup> districts = repository.findAllDistrictGroups(schools(-40L));
 
         assertThat(districts)
                 .usingRecursiveFieldByFieldElementComparator()
