+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/search/exam/student/DefaultStudentExamServiceTest.java
@@ -8,10 +8,9 @@ import org.junit.runner.RunWith;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
 import org.opentestsystem.rdw.reporting.common.model.Student;
+import org.opentestsystem.rdw.reporting.common.security.User;
 import org.opentestsystem.rdw.security.Permission;
 import org.opentestsystem.rdw.security.PermissionScope;
-import org.opentestsystem.rdw.reporting.organization.model.Organization;
-import org.opentestsystem.rdw.reporting.common.security.User;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -72,12 +71,10 @@ public class DefaultStudentExamServiceTest {
         final StudentHistoryExamWrapper wrapper1 = new StudentHistoryExamWrapper.Builder()
                 .assessment(Assessment.builder().id(1).build())
                 .exam(Exam.builder().id(2).build())
-                .school(new Organization(3, "School A"))
                 .build();
         final StudentHistoryExamWrapper wrapper2 = new StudentHistoryExamWrapper.Builder()
                 .assessment(Assessment.builder().id(4).build())
                 .exam(Exam.builder().id(5).build())
-                .school(new Organization(6, "School B"))
                 .build();
 
         final List<StudentHistoryExamWrapper> exams = ImmutableList.of(wrapper1, wrapper2);
