+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/search/exam/student/JdbcStudentExamRepositoryIT.java
@@ -7,10 +7,9 @@ import org.opentestsystem.rdw.reporting.TestData;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
 import org.opentestsystem.rdw.reporting.common.model.Student;
-import org.opentestsystem.rdw.security.PermissionScope;
 import org.opentestsystem.rdw.reporting.common.test.ITDataSourceConfiguration;
 import org.opentestsystem.rdw.reporting.common.test.RepositoryIT;
-import org.opentestsystem.rdw.reporting.organization.model.Organization;
+import org.opentestsystem.rdw.security.PermissionScope;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
@@ -70,11 +69,9 @@ public class JdbcStudentExamRepositoryIT {
         final List<StudentHistoryExamWrapper> wrappers = repository.findAllForStudent(permissions(individualOf(PermissionScope.STATEWIDE)), ImmutableSet.of(), -1L);
         final List<Exam> exams = newArrayList();
         final List<Assessment> assessments = newArrayList();
-        final List<Organization> schools = newArrayList();
         wrappers.forEach(wrapper -> {
             exams.add(wrapper.getExam());
             assessments.add(wrapper.getAssessment());
-            schools.add(wrapper.getSchool());
         });
 
         assertThat(exams)
