+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/search/exam/student/StudentExamControllerIT.java
@@ -5,7 +5,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
 import org.opentestsystem.rdw.reporting.common.model.Exam;
-import org.opentestsystem.rdw.reporting.organization.model.Organization;
+import org.opentestsystem.rdw.reporting.common.model.Organization;
 import org.opentestsystem.rdw.reporting.common.model.Student;
 import org.opentestsystem.rdw.reporting.common.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -47,13 +47,11 @@ public class StudentExamControllerIT {
         final Student student = Student.builder().id(123).build();
         final StudentHistoryExamWrapper wrapper1 = new StudentHistoryExamWrapper.Builder()
                 .assessment(Assessment.builder().id(1).build())
-                .exam(Exam.builder().id(1).build())
-                .school(new Organization(1, "School A"))
+                .exam(Exam.builder().id(1).school(new Organization(1, "school1")).build())
                 .build();
         final StudentHistoryExamWrapper wrapper2 = new StudentHistoryExamWrapper.Builder()
                 .assessment(Assessment.builder().id(2).build())
-                .exam(Exam.builder().id(2).build())
-                .school(new Organization(2, "School B"))
+                .exam(Exam.builder().id(2).school(new Organization(2, "school2")).build())
                 .build();
 
         when(service.getExams(any(User.class), eq(123L))).thenReturn(
@@ -70,8 +68,8 @@ public class StudentExamControllerIT {
                 .andExpect(jsonPath("$.exams.[*].assessment[?(@.id==2)].id").exists())
                 .andExpect(jsonPath("$.exams.[*].exam[?(@.id==1)].id").exists())
                 .andExpect(jsonPath("$.exams.[*].exam[?(@.id==2)].id").exists())
-                .andExpect(jsonPath("$.exams.[*].school[?(@.id==1)].id").exists())
-                .andExpect(jsonPath("$.exams.[*].school[?(@.id==2)].id").exists());
+                .andExpect(jsonPath("$.exams.[*].exam[?(@.id==1)].school.id").exists())
+                .andExpect(jsonPath("$.exams.[*].exam[?(@.id==2)].school.id").exists());
     }
 
     @Test
