+++ webapp/src/test/java/org/opentestsystem/rdw/reporting/search/user/UserContextControllerIT.java
@@ -3,9 +3,9 @@ package org.opentestsystem.rdw.reporting.search.user;
 import com.google.common.collect.ImmutableSet;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.opentestsystem.rdw.reporting.organization.model.School;
-import org.opentestsystem.rdw.reporting.search.group.Group;
+import org.opentestsystem.rdw.reporting.common.model.School;
 import org.opentestsystem.rdw.reporting.common.security.User;
+import org.opentestsystem.rdw.reporting.search.group.Group;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
